+++ js/main.js
@@ -1,1 +1,3 @@
-import './data.js';
+import {photosArray} from './data.js';
+
+console.log(photosArray);
